backend/requirements.txt
Flask==3.0.2
requests==2.31.0
websockets==12.0
pyyaml==6.0.1
